backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose
passlib[bcrypt]
pydantic
python-multipart
pdfplumber
pymupdf
pytesseract
pillow 